Move CollectionOperations.saveNoteType() to CardTemplateNoteType
It was used only there, allows removal of the now empty async package.
